\name{dbeOutputSim}
\docType{class}
\alias{dbeOutputSim}
\title{Class "dbeOutputSim"}
\description{Outcome object from applying \code{COSTdbe} methods to simData class objects. It has the same main 
estructure as dbeOutput class objects but the dataframes generated by the estimation methods include a
variable \code{sample} which identifies the simulated data set.}

\section{Objects from the Class}{
The creator function \code{dbeOutputSim} can be called to create objects from this class.
}

\section{Slots}{
\tabular{lllll}{
\bold{slot} \tab \bold{desc} \tab \bold{elements} \tab \bold{class} \tab \bold{description} \cr
\bold{\code{desc}} \tab \tab \tab \code{character} \tab Descriptive slot \cr
\bold{\code{species}} \tab \tab \tab \code{character} \tab Species \cr
\bold{\code{catchCat}} \tab \tab \tab \code{character} \tab Catch category (eg "LAN", "DIS" or "all") \cr
\bold{\code{param}} \tab \tab \tab \code{character} \tab Parameter estimated (eg "maturity","sex-ratio",...)  \cr
\bold{\code{strataDesc}} \tab \tab \tab \code{strIni} \tab Stratification considered \cr
\bold{\code{methodDesc}} \tab \tab \tab \code{character} \tab Used method (eg "analytical","bootstrap" or "Bayesian") \cr
}
All the following slots contain numerics or dataframes that will be generated by COSTdbe methods :\cr
\tabular{lllll}{
\bold{nSamp} \tab \tab \tab \code{list} \tab Number of samples\cr
 \tab \tab len \tab \code{data.frame} \tab Number of length samples. \cr
 \tab \tab age \tab \code{data.frame} \tab Number of age samples. \cr
\bold{nMes} \tab \tab \tab \code{list} \tab Number of individual measured\cr
 \tab \tab len \tab \code{data.frame} \tab Number of individual measured for length data. \cr
 \tab \tab age \tab \code{data.frame} \tab Number of individual measured for age data. \cr
\bold{lenStruc} \tab \tab \tab \code{list} \tab Estimates of the length structure\cr
 \tab \tab estim \tab \code{data.frame} \tab Final estimates (\emph{length} field added). \cr
 \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{length} and \emph{iter} fields added). \cr
\bold{lenVar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'lenStruc'\cr
\bold{lenNum} \tab \tab \tab \code{list} \tab Precision estimates for length structure\cr
 \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
\bold{ageStruc} \tab \tab \tab \code{list} \tab Estimates of the age structure\cr
 \tab \tab estim \tab \code{data.frame} \tab Final estimates (\emph{age} field added). \cr
 \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{age} and \emph{iter} fields added). \cr
\bold{ageVar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'ageStruc'\cr
\bold{ageNum} \tab \tab \tab \code{list} \tab Precision estimates for age structure\cr
 \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
\bold{totalN} \tab \tab \tab \code{list} \tab Estimates of the total number of the parameters\cr
 \tab \tab estim \tab \code{data.frame} \tab Final estimates. \cr
 \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{iter} field added). \cr
\bold{totalNvar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'totalN'\cr
\bold{totalNnum} \tab \tab \tab \code{list} \tab Precision estimates for total numbers\cr
 \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
\bold{totalW} \tab \tab \tab \code{list} \tab Estimates of the total weight of the parameters\cr
 \tab \tab estim \tab \code{data.frame} \tab Final estimates. \cr
 \tab \tab rep \tab \code{data.frame} \tab Resampling replicates (\emph{iter} field added). \cr
\bold{totalWvar} \tab \tab \tab \code{data.frame} \tab Estimates of the variance of 'totalW'\cr
\bold{totalWnum} \tab \tab \tab \code{list} \tab Precision estimates for total weights\cr
 \tab \tab ci \tab \code{data.frame} \tab Confidence intervals. \cr
 \tab \tab cv \tab \code{data.frame} \tab Coefficients of variation. \cr
 \tab \tab DCRcvIndicator \tab \code{numeric} \tab Weighted global cv. \cr
}}

\author{Dorleta Garcia \email{dgarcia@azti.es}}

\seealso{
\code{\link[COSTdbe]{dbeOutput}}
}

\examples{
showClass("dbeOutputSim")
}

\keyword{classes}

