\name{simSamples}
\alias{simSamples}
\title{ Simulation of data sets }
\description{
  This function simulates data sets for the evaluation and comparison of the estimation methods
}
\usage{
simSamples(obj, ...)
}
\arguments{
  \item{obj}{ a 'simData' object }
  \item{\dots}{ Further arguments }
}

\value{
  It returns a simData object with the \code{sample} slots containing the simulated data sets
}

\author{Dorleta Garcia \email{dgarcia@azti.es}}

\seealso{ 
\code{\link{simData}}
}

%\examples{ }

\keyword{methods}
